requirements.txt
Django==3.2.16
boto3==1.24.86
# dj-database-url==1.0.0
django-dotenv==1.4.2
django-storages==1.13.1
gunicorn==20.1.0
psycopg2-binary==2.9.4
requests==2.28.1
celery==5.2.7
djangorestframework==3.14.0
django-rest-swagger==2.2.0
python-decouple==3.8
django-storages==1.13.2
django-import-export==3.1.0
django-celery-results==2.5.0
django-cors-headers==3.14.0
django-admin-rangefilter==0.10.0
django-multiselectfield==0.1.12
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
django-cleanup==7.0.0
django-widget-tweaks==1.4.12
django-filter==23.1
django-crum==0.7.9